#96 [REFACTOR] : 서버 변동에 따른 수정 및 onClickListener 코드 리팩토링
