+++ app/src/main/java/com/example/charo_android/api/login/SignInViewService.kt
@@ -1,7 +1,7 @@
-package com.example.charo_android.api
+package com.example.charo_android.api.login
 
-import com.example.charo_android.data.RequestSignInData
-import com.example.charo_android.data.ResponseSignInData
+import com.example.charo_android.data.login.RequestSignInData
+import com.example.charo_android.data.login.ResponseSignInData
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.Headers
@@ -10,7 +10,7 @@ import retrofit2.http.POST
 interface SignInViewService {
     @Headers ("accept: application/json",
         "content-type: application/json")
-    @POST("/sign/signIn")
+    @POST("/user/login")
     fun postSignIn(
         @Body body: RequestSignInData
     ): Call<ResponseSignInData>
