--- app/src/main/java/com/example/charo_android/data/RequestSignInData.kt
@@ -1,6 +1,0 @@
-package com.example.charo_android.data
-
-data class RequestSignInData(
-    val id: String,
-    val password: String
-)
