+++ app/src/main/java/com/example/charo_android/data/login/RequestSignInData.kt
@@ -1,0 +1,6 @@
+package com.example.charo_android.data.login
+
+data class RequestSignInData(
+    val email: String,
+    val password: String
+)
